fix: support registries that do not provide referrers API
